+++ src/models/splinter.h
@@ -3,30 +3,30 @@
 
 #include "mat.h"
 
-const float wheel_diameter     = 0.25;
-const float baseline           = 0.35;
-const float torque_constant    = 3.0; // torque (Nm) per amp
-const float emf_constant       = 2.0; // volts per rad_per_sec
-const float winding_resistance = 5.5; // ohms
-const float inertia            = 0.5; // kg*m^2
-const float drag_constant      = 1.0; // drag (Nm per rad_per_sec) ( >= 0)
-const float dt                 = 0.016; // need a better way to figure this out
+const double wheel_diameter     = 0.25;
+const double baseline           = 0.35;
+const double torque_constant    = 3.0; // torque (Nm) per amp
+const double emf_constant       = 2.0; // volts per rad_per_sec
+const double winding_resistance = 5.5; // ohms
+const double inertia            = 0.5; // kg*m^2
+const double drag_constant      = 1.0; // drag (Nm per rad_per_sec) ( >= 0)
+const double dt                 = 0.016; // need a better way to figure this out
 
-inline void calc_rps(float &rps, float input_volts) {
-	float volts_emf = rps * emf_constant;
-	float amps = (input_volts - volts_emf) / winding_resistance;
-	float torque0 = amps * torque_constant;
-	float torque_drag = rps * drag_constant;
-	float torque_net = torque0 - torque_drag;
-	float acceleration = torque_net / inertia;
+inline void calc_rps(double &rps, double input_volts) {
+	double volts_emf = rps * emf_constant;
+	double amps = (input_volts - volts_emf) / winding_resistance;
+	double torque0 = amps * torque_constant;
+	double torque_drag = rps * drag_constant;
+	double torque_net = torque0 - torque_drag;
+	double acceleration = torque_net / inertia;
 	rps += acceleration * dt;
 }
 
-inline void splinter_update(float &px, float &py, float &vx, float &vy, float &rz, float &rtz, float &lrps, float &rrps, float lvolt, float rvolt) {
+inline void splinter_update(double &px, double &py, double &vx, double &vy, double &rz, double &rtz, double &lrps, double &rrps, double lvolt, double rvolt) {
 	calc_rps(lrps, lvolt * 12);
 	calc_rps(rrps, rvolt * 12);
-	float dleft  = dt * lrps * wheel_diameter;
-	float dright = dt * rrps * wheel_diameter;
+	double dleft  = dt * lrps * wheel_diameter;
+	double dright = dt * rrps * wheel_diameter;
 	
 	Eigen::Matrix3d orient(Eigen::AngleAxis<double>(rz, vec3(0, 0, 1)));
 	vec3 vel = orient * vec3((dleft + dright) / 2, 0., 0.);
